SNOW-1374896: Add parameterized tests for getString and getBytes of structured types
